% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{bayes_cp_test}
\alias{bayes_cp_test}
\title{Bayesian change point analysis on sedimentation rates}
\usage{
bayes_cp_test(dates_1989)
}
\description{
Bayesian change point analysis on sedimentation rates
}
\examples{
\dontrun{
bayes_cp_result <- bayes_cp_test(chrono_data_slopes_plot$dates_1989)
}
}

