% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{bayes_slopes_chrono_data}
\alias{bayes_slopes_chrono_data}
\title{Posteriors of slopes of OSL and C14 depth-age lines}
\usage{
bayes_slopes_chrono_data(upper)
}
\description{
Posteriors of slopes of OSL and C14 depth-age lines
}
\examples{
\dontrun{
bayes_values <- bayes_slopes_chrono_data(ancova_p_values$upper)
}
}

