% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{lens_differences_raw}
\alias{lens_differences_raw}
\title{Differences in assemblage raw materials above and below the pit lens}
\usage{
lens_differences_raw()
}
\description{
Differences in assemblage raw materials above and below the pit lens
}
\examples{
\dontrun{
lens_differences_raw_plot <- lens_differences_raw()
}
}

