% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{lens_differences_tech}
\alias{lens_differences_tech}
\title{Differences in assemblage technology above and below the pit lens}
\usage{
lens_differences_tech()
}
\description{
Differences in assemblage technology above and below the pit lens
}
\examples{
\dontrun{
lens_differences_tech_plot <- lens_differences_tech()
}
}

